refac: App 컴포넌트 단순화
